requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
numpy==2.2.1
requests==2.32.3
google-genai>=0.3.0
python-dotenv==1.0.0
httpx>=0.28.1,<1.0.0

